use same addComputedColumns / applyChanges / reorderColumns as new chart
