+++ lib/dw/chart.js
@@ -9,7 +9,9 @@ import delimited from './dataset/delimited';
 import { name } from './utils';
 import events from './utils/events';
 import { loadScript } from '@datawrapper/shared/fetch';
-import { Parser } from './utils/parser';
+import reorderColumns from './dataset/reorderColumns';
+import applyChanges from './dataset/applyChanges';
+import addComputedColumns from './dataset/addComputedColumns';
 
 let visualization;
 
@@ -29,6 +31,10 @@ export default function(attributes) {
         // returns an attribute
         get(key, _default) {
             return get(attributes, key, _default);
+        },
+
+        getMetadata(key, _default) {
+            return get(attributes, `metadata.${key}`, _default);
         },
 
         set(key, value) {
@@ -68,7 +74,7 @@ export default function(attributes) {
                 if (ds !== true) _ds = ds;
                 dataset = chart.get('metadata.data.json')
                     ? _ds
-                    : reorderColumns(applyChanges(addComputedColumns(_ds)));
+                    : reorderColumns(chart, applyChanges(chart, addComputedColumns(chart, _ds)));
                 if (ds === true) return dataset;
                 return chart;
             }
@@ -226,232 +232,6 @@ export default function(attributes) {
             return changed;
         }
     };
-
-    function reorderColumns(dataset) {
-        const order = chart.get('metadata.data.column-order', []);
-        if (order.length && order.length === dataset.numColumns()) {
-            dataset.columnOrder(order);
-        }
-        return dataset;
-    }
-
-    function applyChanges(dataset) {
-        const changes = chart.get('metadata.data.changes', []);
-        const transpose = chart.get('metadata.data.transpose', false);
-        // apply changes
-        changes.forEach(change => {
-            let row = 'row';
-            let column = 'column';
-            if (transpose) {
-                row = 'column';
-                column = 'row';
-            }
-
-            if (dataset.hasColumn(change[column])) {
-                if (change[row] === 0) {
-                    if (change.previous && change.previous !== 'undefined') {
-                        var oldTitle = dataset.column(change[column]).title();
-                        if (oldTitle !== change.previous) return;
-                    }
-                    dataset.column(change[column]).title(change.value);
-                } else {
-                    if (change.previous && change.previous !== 'undefined') {
-                        var curValue = dataset.column(change[column]).raw(change[row] - 1);
-                        if (curValue !== change.previous) return;
-                    }
-                    dataset.column(change[column]).raw(change[row] - 1, change.value);
-                }
-            }
-        });
-
-        // overwrite column types
-        const columnFormats = chart.get('metadata.data.column-format', {});
-        _.each(columnFormats, function(columnFormat, key) {
-            if (columnFormat.type && dataset.hasColumn(key) && columnFormat.type !== 'auto') {
-                dataset.column(key).type(columnFormat.type);
-            }
-            if (columnFormat['input-format'] && dataset.hasColumn(key)) {
-                dataset
-                    .column(key)
-                    .type(true)
-                    .format(columnFormat['input-format']);
-            }
-        });
-        return dataset;
-    }
-
-    function addComputedColumns(dataset) {
-        const virtualColumns = chart.get('metadata.describe.computed-columns', {});
-        const data = dataset.list();
-        const columnNameToVar = {};
-        const colAggregates = {};
-        const parser = new Parser();
-
-        dataset.eachColumn(function(col) {
-            if (col.isComputed) return;
-            columnNameToVar[col.name()] = columnNameToVariable(col.name());
-            if (col.type() === 'number') {
-                const [min, max] = col.range();
-                colAggregates[col.name()] = {
-                    min,
-                    max,
-                    sum: sum(col.values()),
-                    mean: mean(col.values()),
-                    median: median(col.values())
-                };
-            } else if (col.type() === 'date') {
-                const [min, max] = col.range();
-                colAggregates[col.name()] = { min, max };
-            }
-        });
-
-        _.each(virtualColumns, addComputedColumns);
-
-        return dataset;
-
-        function addComputedColumns(formula, name) {
-            const datefmt = function(d) {
-                return (
-                    d.getFullYear() +
-                    '-' +
-                    leftPad(1 + d.getMonth(), 2, 0) +
-                    '-' +
-                    leftPad(1 + d.getDate(), 2, 0)
-                );
-            };
-            const errors = [];
-            let expr;
-            try {
-                expr = formula.trim() ? parser.parse(formula) : null;
-            } catch (e) {
-                errors.push(e.message);
-            }
-            if (!expr)
-                expr = {
-                    evaluate() {
-                        return '';
-                    }
-                };
-            const values = data
-                .map(function(row, index) {
-                    const context = {
-                        ROWNUMBER: index
-                    };
-                    _.each(row, function(val, key) {
-                        if (!columnNameToVar[key]) return;
-                        context[columnNameToVar[key]] = val;
-                        if (colAggregates[key]) {
-                            Object.keys(colAggregates[key]).forEach(aggr => {
-                                context[`${columnNameToVar[key]}__${aggr}`] =
-                                    colAggregates[key][aggr];
-                            });
-                        }
-                    });
-                    try {
-                        return expr.evaluate(context);
-                    } catch (error) {
-                        errors.push(error.message);
-                        return null;
-                    }
-                })
-                .map(function(v) {
-                    if (_.isBoolean(v)) return v ? 'yes' : 'no';
-                    if (_.isDate(v)) return datefmt(v);
-                    if (_.isNumber(v)) return '' + v;
-                    if (_.isNull(v)) return null;
-                    return String(v);
-                });
-            var virtualColumn = column(name, values);
-            virtualColumn.isComputed = true;
-            virtualColumn.errors = errors;
-            dataset.add(virtualColumn);
-        }
-
-        // some d3 stuff
-        function min(array) {
-            var i = -1;
-            var n = array.length;
-            var a, b;
-            if (arguments.length === 1) {
-                while (++i < n)
-                    // eslint-disable-next-line
-                    if ((b = array[i]) != null && b >= b) {
-                        a = b;
-                        break;
-                    }
-                while (++i < n) if ((b = array[i]) != null && a > b) a = b;
-            }
-            return a;
-        }
-        function max(array) {
-            var i = -1;
-            var n = array.length;
-            var a, b;
-            if (arguments.length === 1) {
-                while (++i < n)
-                    // eslint-disable-next-line
-                    if ((b = array[i]) != null && b >= b) {
-                        a = b;
-                        break;
-                    }
-                while (++i < n) if ((b = array[i]) != null && b > a) a = b;
-            }
-            return a;
-        }
-        function sum(array) {
-            var s = 0;
-            var n = array.length;
-            var a;
-            var i = -1;
-            if (arguments.length === 1) {
-                while (++i < n) if (numeric((a = +array[i]))) s += a;
-            }
-            return s;
-        }
-        function mean(array) {
-            var s = 0;
-            var n = array.length;
-            var a;
-            var i = -1;
-            var j = n;
-            while (++i < n)
-                if (numeric((a = number(array[i])))) s += a;
-                else --j;
-            if (j) return s / j;
-        }
-        function median(array) {
-            var numbers = [];
-            var n = array.length;
-            var a;
-            var i = -1;
-            if (arguments.length === 1) {
-                while (++i < n) if (numeric((a = number(array[i])))) numbers.push(a);
-            }
-            if (numbers.length) return quantile(numbers.sort(ascending), 0.5);
-        }
-        function quantile(values, p) {
-            var H = (values.length - 1) * p + 1;
-            var h = Math.floor(H);
-            var v = +values[h - 1];
-            var e = H - h;
-            return e ? v + e * (values[h] - v) : v;
-        }
-        function number(x) {
-            return x === null ? NaN : +x;
-        }
-        function numeric(x) {
-            return !isNaN(x);
-        }
-        function ascending(a, b) {
-            return a < b ? -1 : a > b ? 1 : a >= b ? 0 : NaN;
-        }
-
-        function leftPad(s, l, pad) {
-            s = String(s);
-            while (s.length < l) s = String(pad) + s;
-            return s;
-        }
-    }
 
     return chart;
 }
